build: turn on the warning flags a modernization pass needs (B1)
-Wall was the only warning flag, and CMakeLists claimed cleanliness under
-Wstrict-aliasing=2, a flag set nowhere. Now -Wextra -Wshadow
-Wnon-virtual-dtor -Wstrict-aliasing=2, plus an opt-in WERROR option (off
by default so a toolchain with a new diagnostic cannot break a build; CI
and this branch should set it).

Third-party includes move to include_directories(SYSTEM ...). That was the
blocker: Boost 1.76 alone emits ~12k -Wold-style-cast and 33
-Wnon-virtual-dtor, which would have forced those warnings off globally
and hidden ours with them. With Boost quiet the whole tree produced 113.

Two of those were real:

  rtmp_protocol still took and stored the chunk size as uint16 -- the
  other half of M21, which widened it on the connection but not here, so
  a Set Chunk Size above 65535 still truncated on the serialize path.

  service::handle_iikeying passed cert_len, a 64-bit VLU off the wire,
  into a uint32 parameter.

The rest were an address_v4 shadowing a live `item` two scopes up, the
AMF3 VLU byte extraction (truncation intended, now explicit), a size_t
length into write_script, m_invoke_id declared long while every id it
feeds is uint32, and nine unused parameters.

A -Werror build with the full set is clean.

+++ amf3.cpp
@@ -129,38 +129,38 @@ namespace fms
 		std::uint8_t b;
 		if ((value & 0xffffff80) == 0)
 		{
-			b = value & 0x7f;
+			b = static_cast<std::uint8_t>(value & 0x7f);
 			buffer << b;
 			return;
 		}
 
 		if ((value & 0xffffc000) == 0)
 		{
-			b = (value >> 7) | 0x80;
+			b = static_cast<std::uint8_t>((value >> 7) | 0x80);
 			buffer << b;
-			b = value & 0x7f;
+			b = static_cast<std::uint8_t>(value & 0x7f);
 			buffer << b;
 			return;
 		}
 
 		if ((value & 0xffe00000) == 0)
 		{
-			b = (value >> 14) | 0x80;
+			b = static_cast<std::uint8_t>((value >> 14) | 0x80);
 			buffer << b;
-			b = (value >> 7) | 0x80;
+			b = static_cast<std::uint8_t>((value >> 7) | 0x80);
 			buffer << b;
-			b = value & 0x7f;
+			b = static_cast<std::uint8_t>(value & 0x7f);
 			buffer << b;
 			return;
 		}
 
-		b = (value >> 22) | 0x80;
+		b = static_cast<std::uint8_t>((value >> 22) | 0x80);
 		buffer << b;
-		b = (value >> 15) | 0x80;
+		b = static_cast<std::uint8_t>((value >> 15) | 0x80);
 		buffer << b;
-		b = (value >> 8) | 0x80;
+		b = static_cast<std::uint8_t>((value >> 8) | 0x80);
 		buffer << b;
-		b = value & 0xff;
+		b = static_cast<std::uint8_t>(value & 0xff);
 		buffer << b;
 	}
